Add some util methods
